DBZ-8956 refactor native integration test
