+++ quarkus-debezium-parent/quarkus-debezium-postgres-parent/integration-test/src/main/resources/application.properties
@@ -19,6 +19,9 @@ quarkus.debezium.configuration.predicates.p2.type=org.apache.kafka.connect.trans
 quarkus.debezium.configuration.predicates=p2
 
 
-# Agroal config
-quarkus.datasource.db-kind=postgresql
-quarkus.datasource.devservices.image-name=quay.io/debezium/postgres:15
+quarkus.datasource.devservices.enabled=false
+quarkus.debezium.configuration.database.hostname=localhost
+quarkus.debezium.configuration.database.port=5432
+quarkus.debezium.configuration.database.user=native
+quarkus.debezium.configuration.database.password=native
+quarkus.debezium.configuration.database.dbname=native
