--- quarkus-debezium-parent/quarkus-debezium-postgres-parent/integration-test/src/test/resources/application-native-it.properties
@@ -1,37 +1,0 @@
-#
-# /*
-#  * Copyright Debezium Authors.
-#  *
-#  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
-#  */
-#
-
-# Debezium CDC configuration
-quarkus.debezium.configuration.offset.storage=org.apache.kafka.connect.storage.MemoryOffsetBackingStore
-quarkus.debezium.configuration.name=native
-quarkus.debezium.configuration.topic.prefix=dbserver1
-quarkus.debezium.configuration.table.include.list=inventory.products
-quarkus.debezium.configuration.plugin.name=pgoutput
-quarkus.debezium.configuration.snapshot.mode=never
-
-
-# Transformation
-quarkus.debezium.configuration.transforms.t0.add.fields=op,table
-quarkus.debezium.configuration.transforms.t0.add.headers=db,table
-quarkus.debezium.configuration.transforms.t0.negate=false
-quarkus.debezium.configuration.transforms.t0.predicate=p2
-quarkus.debezium.configuration.transforms.t0.type=io.debezium.transforms.ExtractNewRecordState
-quarkus.debezium.configuration.transforms=t0
-quarkus.debezium.configuration.predicates.p2.pattern=inventory.inventory.products
-quarkus.debezium.configuration.predicates.p2.type=org.apache.kafka.connect.transforms.predicates.TopicNameMatches
-quarkus.debezium.configuration.predicates=p2
-
-# Quarkus
-quarkus.native.resources.includes=META-INF/services/*
-quarkus.native.native-image-agent.enabled=true
-quarkus.native.additional-build-args=--initialize-at-run-time=org.apache.kafka.common.security.authenticator.SaslClientAuthenticator
-
-
-# Agroal config
-quarkus.datasource.db-kind=postgresql
-quarkus.datasource.devservices.image-name=quay.io/debezium/postgres:15
